feat: implement footer component and update layout structure
